[Kernels] fix windows fail (#542)
* fix windows fail

* disable jit kernel check in windows

* add TODO for reference implementation

* rename transpose matmul UT
