+++ intel_extension_for_transformers/backends/neural_engine/kernels/CMakeLists.txt
@@ -35,8 +35,10 @@ file(GLOB HEADERS
 
 if (NE_DYNAMIC_LINK)
     add_library(${HOST_LIBRARY_NAME} SHARED ${SOURCES} ${HEADERS})
+    target_compile_definitions(${HOST_LIBRARY_NAME} PRIVATE SPARSE_KERNEL_BUILD)
 else()
     add_library(${HOST_LIBRARY_NAME} STATIC ${SOURCES} ${HEADERS})
+    target_compile_definitions(${HOST_LIBRARY_NAME} PRIVATE SPARSE_API_)
 endif()
 
 install(TARGETS ${HOST_LIBRARY_NAME})
