+++ intel_extension_for_transformers/backends/neural_engine/kernels/include/kernels/matmul_ref.hpp
@@ -32,10 +32,12 @@ namespace jd {
 //   3. jit_xxxx_t is JIT assembly implementation of a specific derived primitive/kernel.
 //   where, "xxxx" represents an algorithm, such as brgemm, GEMM and so on.
 class matmul_ref_k_t;
+
+// TODO(hengyu): hide reference implementation from users
 /**
  * @brief a derived kernel descriptor. ref_param_t is its class member.
  */
-class matmul_ref_kd_t : public kernel_desc_t {
+class SPARSE_API_ matmul_ref_kd_t : public kernel_desc_t {
  public:
   explicit matmul_ref_kd_t(const jd::operator_desc& op_desc)
       : kernel_desc_t(kernel_kind::sparse_matmul), op_desc_(op_desc), shape_(5) {}
@@ -65,7 +67,7 @@ class matmul_ref_kd_t : public kernel_desc_t {
 /**
  * @brief a derived kernel. kd_t and jit_domain are its class members.
  */
-class matmul_ref_k_t : public kernel_t {
+class SPARSE_API_ matmul_ref_k_t : public kernel_t {
  public:
   using kd_t = matmul_ref_kd_t;
   explicit matmul_ref_k_t(const std::shared_ptr<const kd_t>& kd);
