+++ intel_extension_for_transformers/backends/neural_engine/test/gtest/kernels/CMakeLists.txt
@@ -42,6 +42,9 @@ list(APPEND KERNEL_TEST_CASES_SRC test_attention.cpp)
 endif()
 if(UNIX)
 list(APPEND KERNEL_TEST_CASES_SRC test_spmm_amx_bf16_x16_kernel.cpp)
+list(APPEND KERNEL_TEST_CASES_SRC test_matmul_vnni_p2031_p2013_tr_src0.cpp)
+list(APPEND KERNEL_TEST_CASES_SRC test_matmul_vnni_p2031_p2013_tr_src1.cpp)
+list(APPEND KERNEL_TEST_CASES_SRC test_matmul_vnni_p2031_p2013_mm_tile.cpp)
 endif()
 
 function(register_gtest_func exe src)
