+++ intel_extension_for_transformers/backends/neural_engine/test/gtest/kernels/test_matmul_vnni_p2031_p2013.cpp
@@ -24,8 +24,6 @@
 #include "unit_test_utils.hpp"
 #include "kernels/matmul_types.hpp"
 #include "kernels/matmul_ref.hpp"
-#include "jit_domain/jit_seq_cpy_2x8x8.hpp"
-#include "jit_domain/jit_seq_cpy_48x4.hpp"
 #include "jit_domain/jit_matmul_vnni_8xkx48.hpp"
 
 #define OMP_NUM_THREADS "OMP_NUM_THREADS"
@@ -33,344 +31,6 @@
 namespace jd {
 using dt = jd::data_type;
 using ft = jd::format_type;
-
-// test seq_cpy_2x8x8
-struct seqcpyA_param_t {
-  dim_t M;
-  dim_t N;
-  uint8_t offset;
-};
-inline static std::string seqcpyATestParam2str(testing::TestParamInfo<seqcpyA_param_t> tpi) {
-  return std::to_string(tpi.param.M) + "_" + std::to_string(tpi.param.N) + "_off" + std::to_string(tpi.param.offset);
-}
-class MMVNNIP2031P2013SEQCPYATest : public testing::TestWithParam<seqcpyA_param_t> {
- protected:
-  MMVNNIP2031P2013SEQCPYATest() {}
-  virtual ~MMVNNIP2031P2013SEQCPYATest() {}
-  void SetUp() override {}
-  void TearDown() override {}
-};
-TEST_P(MMVNNIP2031P2013SEQCPYATest, ) {
-  seqcpyA_param_t t = testing::TestWithParam<seqcpyA_param_t>::GetParam();
-  const int M = t.M, N = t.N;
-  const int ld_src = N;
-  if (M % 8 != 0) {
-    SPARSE_LOG(WARNING) << "M must be a multiple of 8";
-    return;
-  }
-  size_t dst_size = M * (ceil_div(N, 8) * 8);
-  int8_t* src = new int8_t[M * N];
-  for (auto i = 0; i < M * N; ++i) src[i] = i % UINT8_MAX - INT8_MIN;
-  uint8_t* dst = reinterpret_cast<uint8_t*>(aligned_alloc(64, dst_size));
-  memset(dst, 3, dst_size);
-
-  // get true data
-  uint8_t* dst_ref = new uint8_t[dst_size];
-  memset(dst_ref, 3, dst_size);
-  for (auto j = 0; j < N; j += 8)
-    for (auto i = 0; i < M; i += 4)
-      for (auto ii = 0; ii < 4; ++ii)
-        for (auto jj = 0; jj < 8; ++jj)
-          dst_ref[j * M + i * 8 + jj * 4 + ii] =
-              (i + ii >= M || j + jj >= N) ? 0 : (src[(i + ii) * ld_src + j + jj] + t.offset);
-
-  // run kernel
-  auto jit_ker = new jit_seq_cpy_2x8x8({M, N, ld_src, t.offset});
-  ASSERT_NE(jit_ker, nullptr);
-  ASSERT_TRUE(jit_ker->create_kernel());
-  for (dim_t i = 0; i < M; i += 8) {
-    jit_seq_cpy_2x8x8::rt_data_t rt_data{
-        src + i * N,
-        dst + i * 8,
-    };
-    (*jit_ker)(&rt_data);
-  }
-
-  ASSERT_TRUE(compare_data<uint8_t>(dst, dst_size, dst_ref, dst_size, 0));
-
-  delete[] src;
-  delete[] dst_ref;
-  free(dst);
-  delete jit_ker;
-}
-INSTANTIATE_TEST_SUITE_P(SparseLib, MMVNNIP2031P2013SEQCPYATest,
-                         ::testing::ValuesIn(std::vector<seqcpyA_param_t>{
-                             {64, 64, 0},
-                             {64, 1, 0},
-                             {64, 15, 0},
-                             {64, 63, 0},
-                             {64, 65, 0},
-                             {64, 64, 128},
-                             {64, 1, 128},
-                             {64, 15, 128},
-                             {64, 63, 128},
-                             {64, 65, 131},
-                         }),
-                         seqcpyATestParam2str);
-
-// test seq_cpy_48x4
-
-struct seqcpyB_param_t {
-  dim_t M;
-  dim_t N;
-  bool sum;
-};
-inline static std::string seqcpyBTestParam2str(testing::TestParamInfo<seqcpyB_param_t> tpi) {
-  std::string repr = std::to_string(tpi.param.M) + "_" + std::to_string(tpi.param.N);
-  if (tpi.param.sum) repr += "_sum";
-  return repr;
-}
-class MMVNNIP2031P2013SEQCPYBTest : public testing::TestWithParam<seqcpyB_param_t> {
- protected:
-  MMVNNIP2031P2013SEQCPYBTest() {}
-  virtual ~MMVNNIP2031P2013SEQCPYBTest() {}
-  void SetUp() override {}
-  void TearDown() override {}
-};
-TEST_P(MMVNNIP2031P2013SEQCPYBTest, ) {
-  seqcpyB_param_t t = testing::TestWithParam<seqcpyB_param_t>::GetParam();
-  const int M = t.M, N = t.N;
-  const int ld_src = N;
-  const int pad_n = pad_to(N, 48);
-  const int max_sum_size = pad_n * 16;  // up to 16 threads
-  if (M % 4 != 0) {
-    SPARSE_LOG(WARNING) << "M must be a multiple of 4! Test case skipped.";
-    return;
-  }
-  const size_t dst_size = M * (ceil_div(N, 48) * 48);
-  uint8_t* const src = new uint8_t[M * N];
-  for (auto i = 0; i < M * N; ++i) src[i] = i % UINT8_MAX;
-
-  // get true data
-  uint8_t* const dst_ref = new uint8_t[dst_size];
-  memset(dst_ref, 0, dst_size);
-#pragma omp parallel for collapse(3)
-  for (auto j = 0; j < N; j += 48)
-    for (auto i = 0; i < M; i += 4)
-      for (auto ii = 0; ii < 4; ++ii)
-#pragma omp simd
-        for (auto jj = 0; jj < 48; ++jj) {
-          auto value = (i + ii >= M || j + jj >= N) ? 0 : src[(i + ii) * ld_src + j + jj];
-          dst_ref[j * M + i * 48 + jj * 4 + ii] = value;
-        }
-
-  // run kernel
-  const n_thread_t with_n_thread(3);
-  uint8_t* const dst = aligned_allocator_t<uint8_t>::allocate(dst_size, true);
-  int32_t* const dst_sum = aligned_allocator_t<int32_t>::allocate(max_sum_size);
-  memset(dst_sum, -1, max_sum_size * sizeof(int32_t));
-  auto jit_ker = new jit_seq_cpy_48x4({M, N, ld_src, t.sum, true});
-  ASSERT_NE(jit_ker, nullptr);
-  ASSERT_TRUE(jit_ker->create_kernel());
-  int real_num_threads = 0;
-  bool sum_append = false;
-#pragma omp parallel for firstprivate(sum_append)
-  for (dim_t i = 0; i < M; i += 4) {
-    jit_seq_cpy_48x4 ::rt_data_t rt_data{
-        src + i * N,
-        dst + i * 48,
-        dst_sum + pad_n * omp_get_thread_num(),
-        sum_append,
-    };
-    (*jit_ker)(&rt_data);
-    sum_append = true;
-    if (i == 0) real_num_threads = omp_get_num_threads();
-  }
-  real_num_threads = std::min(real_num_threads, ceil_div(M, 4));
-
-  ASSERT_TRUE(compare_data<uint8_t>(dst, dst_size, dst_ref, dst_size, 0));
-
-  if (t.sum) {
-    // collaps sum
-    for (dim_t i = 1; i < real_num_threads; ++i) {
-#pragma omp simd
-      for (dim_t j = 0; j < N; ++j) {
-        dst_sum[j] += dst_sum[i * pad_n + j];
-      }
-    }
-
-    //  sum true data
-    int32_t* dst_sum_ref = aligned_allocator_t<int32_t>::allocate(pad_n, true);
-#pragma omp parallel for
-    for (auto j = 0; j < N; ++j)
-      for (auto i = 0; i < M; ++i) {
-        dst_sum_ref[j] += src[i * N + j];
-      }
-
-    ASSERT_TRUE(compare_data<int32_t>(dst_sum, N, dst_sum_ref, N, 0));
-    aligned_allocator_t<int32_t>::deallocate(dst_sum_ref);
-  }
-
-  delete[] src;
-  delete[] dst_ref;
-  aligned_allocator_t<uint8_t>::deallocate(dst);
-  aligned_allocator_t<int32_t>::deallocate(dst_sum);
-  delete jit_ker;
-}
-INSTANTIATE_TEST_SUITE_P(SparseLib, MMVNNIP2031P2013SEQCPYBTest,
-                         ::testing::ValuesIn(std::vector<seqcpyB_param_t>{
-                             {4, 48},
-                             {128, 768, true},
-                             {64, 48, true},
-                             {4, 1, true},
-                             {64, 47, true},
-                             {64, 49, true},
-                             {64, 65, true},
-                         }),
-                         seqcpyBTestParam2str);
-
-// test mm_8xkx48
-struct mm_param_t {
-  dim_t K;
-  dim_t ld_dst;
-  uint8_t bias_lshift;
-  bool binary_add;
-  std::vector<postop_attr> postop_list;
-};
-
-inline static std::string mmTestParam2str(testing::TestParamInfo<mm_param_t> tpi) {
-  std::vector<std::string> params;
-  params.push_back(std::to_string(tpi.param.K));
-  params.push_back(std::to_string(tpi.param.ld_dst));
-  params.push_back("lsh" + std::to_string(tpi.param.bias_lshift));
-  if (tpi.param.binary_add) params.push_back("badd");
-  for (postop_attr& p_attr : tpi.param.postop_list) {
-    params.push_back("post" + std::string(postop_alg_name.at(p_attr.op_alg)));
-    params.push_back(std::string(data_type_name.at(p_attr.dt)));
-    params.push_back(num2id(p_attr.alpha));
-    params.push_back(num2id(p_attr.beta));
-    params.push_back(num2id(p_attr.scale));
-  }
-  return join_str(params, "_");
-}
-class MMVNNIP2031P2013MMTileTest : public testing::TestWithParam<mm_param_t> {
- protected:
-  MMVNNIP2031P2013MMTileTest() {}
-  virtual ~MMVNNIP2031P2013MMTileTest() {}
-  void SetUp() override {}
-  void TearDown() override {}
-};
-
-template <typename dt_dst>
-inline void exec_mmtile(jit_matmul_vnni_8xkx48_t* ker, const uint8_t* src0, const int8_t* src1, const int32_t* bias,
-                        const float* src_b0, dt_dst* dst) {
-  jit_matmul_vnni_8xkx48_t::rt_data_t<dt_dst> rt_data;
-  rt_data.src0 = src0;
-  rt_data.src1 = src1;
-  rt_data.bias = bias;
-  rt_data.src_b0 = src_b0;
-  rt_data.dst = dst;
-  (*ker)(&rt_data);
-}
-
-TEST_P(MMVNNIP2031P2013MMTileTest, ) {
-  const mm_param_t t = testing::TestWithParam<mm_param_t>::GetParam();
-  const data_type dst_type = t.postop_list.size() != 0 ? t.postop_list.back().dt : data_type::fp32;
-  const dim_t M = 8, K = t.K, N = 48;
-  const size_t dst_buf_size = M * t.ld_dst;
-  const size_t dst_buf_bytes = dst_buf_size * type_size.at(dst_type);
-  SPARSE_LOG_IF(FATAL, t.ld_dst < N) << "ld_dst must be gretaer then N";
-  uint8_t* const src0 = new uint8_t[M * K];
-  init_vector(src0, M * K, 0, 10);
-  int8_t* const src1 = new int8_t[K * N];
-  init_vector(src1, N * K, -10, 10);
-  float* src_b0 = nullptr;
-  if (t.binary_add) {
-    src_b0 = new float[dst_buf_size];
-    init_vector(src_b0, dst_buf_size, -10, 10);
-  }
-
-  int32_t* const bias = new int32_t[N];
-  init_vector(bias, N, 0, 100);
-  float scale[1];
-  init_vector(scale, 1);
-
-  // get true data
-  void* const dst0_ref = new char[dst_buf_bytes];
-  memset(dst0_ref, 3, dst_buf_bytes);  // something special to check if the kernel keep the additional memory untouched
-  const auto dst0_ref_u8 = reinterpret_cast<uint8_t*>(dst0_ref);
-  const auto dst0_ref_s8 = reinterpret_cast<int8_t*>(dst0_ref);
-  const auto dst0_ref_f32 = reinterpret_cast<float*>(dst0_ref);
-  for (int i = 0; i < M; ++i) {
-    for (int j = 0; j < N; ++j) {
-      float value = -(bias[j] << t.bias_lshift);
-      for (int k = 0; k < K; ++k) {
-        float l = src0[i * 4 + k % 4 + (k / 4) * 4 * M];
-        float r = src1[j * 4 + k % 4 + (k / 4) * 4 * N];
-        value += l * r;
-      }
-      value *= scale[0];
-      if (t.binary_add) value += src_b0[i * t.ld_dst + j];
-      value = apply_postop_list(value, t.postop_list);
-      switch (dst_type) {
-        case data_type::u8:
-          dst0_ref_u8[i * t.ld_dst + j] = static_cast<uint8_t>(value);
-          break;
-        case data_type::s8:
-          dst0_ref_s8[i * t.ld_dst + j] = static_cast<int8_t>(value);
-          break;
-        case data_type::fp32:
-          dst0_ref_f32[i * t.ld_dst + j] = value;
-          break;
-        default:
-          SPARSE_LOG(FATAL) << "Unexpected dst type!";
-          break;
-      }
-    }
-  }
-
-  // run kernel
-  void* const dst0 = new char[dst_buf_bytes];
-  memset(dst0, 3, dst_buf_bytes);  // something special to check if the kernel keep the additional memory untouched
-  auto jit_ker = new jit_matmul_vnni_8xkx48_t(jit_matmul_vnni_8xkx48_t::param_t{
-      t.K,
-      t.ld_dst,
-      scale[0],
-      t.bias_lshift,
-      t.binary_add,
-      dst_type,
-      t.postop_list,
-      M,
-      N,
-  });
-  ASSERT_NE(jit_ker, nullptr);
-  ASSERT_TRUE(jit_ker->create_kernel());
-
-  switch (dst_type) {
-    case data_type::u8:
-      exec_mmtile<uint8_t>(jit_ker, src0, src1, bias, src_b0, reinterpret_cast<uint8_t*>(dst0));
-      ASSERT_TRUE(compare_data<uint8_t>(dst0, dst_buf_size, dst0_ref, dst_buf_size, 8e-3));
-      break;
-    case data_type::s8:
-      exec_mmtile<int8_t>(jit_ker, src0, src1, bias, src_b0, reinterpret_cast<int8_t*>(dst0));
-      ASSERT_TRUE(compare_data<int8_t>(dst0, dst_buf_size, dst0_ref, dst_buf_size, 8e-3));
-      break;
-    case data_type::fp32:
-      exec_mmtile<float>(jit_ker, src0, src1, bias, src_b0, reinterpret_cast<float*>(dst0));
-      ASSERT_TRUE(compare_data<float>(dst0, dst_buf_size, dst0_ref, dst_buf_size, 8e-3));
-      break;
-    default:
-      SPARSE_LOG(FATAL) << "Unexpected dst type!";
-      break;
-  }
-  delete[] src0;
-  delete[] src1;
-  if (src_b0 != nullptr) delete[] src_b0;
-  delete[] bias;
-  delete[] reinterpret_cast<char*>(dst0);
-  delete[] reinterpret_cast<char*>(dst0_ref);
-  delete jit_ker;
-}
-INSTANTIATE_TEST_SUITE_P(
-    SparseLib, MMVNNIP2031P2013MMTileTest,
-    ::testing::ValuesIn(std::vector<mm_param_t>{
-        {16, 48, 2, false, {}},                                                                   // basic
-        {64, 64, 7, true, {}},                                                                    // ld_dst != N
-        {64, 65, 7, true, {}},                                                                    // ld_dst != N
-        {64, 64, 7, true, {{dt::u8, postop_type::eltwise, postop_alg::quantize, 0.0, 0.0, 10}}},  // post quant s8
-        {64, 64, 7, true, {{dt::u8, postop_type::eltwise, postop_alg::quantize, 128, 0.0, 10}}},  // post quant u8
-    }),
-    mmTestParam2str);
 
 // test trmm vnni
 struct op_args_t {
